renames challenge.ts -> challenge_schema
