+++ specification/challenge_schema.ts
